289: converted test from Xtend to Java
Task-Url: http://github.com/LorenzoBettini/edelta/issues/289
